+++ includes/layouts/settings.php
@@ -95,11 +95,6 @@
 
 				<form method="post" action="">
 					<?php
-					$disble_featured_image = '';
-					if ( isset( $settings['general']['rss-feeds'] ) && 1 === intval( $settings['general']['rss-feeds'] ) ) {
-						$disble_featured_image = 'checked';
-					}
-
 					$disble_default_style = 0;
 					if ( isset( $settings['general']['disable-default-style'] ) && 1 === intval( $settings['general']['disable-default-style'] ) ) {
 						$disble_default_style = 1;
@@ -107,19 +102,15 @@
 
 					$feedzy_delete_days = isset( $settings['general']['feedzy-delete-days'] ) ? $settings['general']['feedzy-delete-days'] : 0;
 					$default_thumbnail_id = isset( $settings['general']['default-thumbnail-id'] ) ? $settings['general']['default-thumbnail-id'] : 0;
+					$telemetry_enabled = get_option( 'feedzy_rss_feeds_logger_flag', 0 );
+
 					switch ( $active_tab ) {
 						case 'general':
 							?>
 							<div class="fz-form-wrap">
 								<div class="form-block">
 									<div class="fz-form-group">
-										<div class="fz-form-switch pb-16">
-											<input type="checkbox" id="rss-feeds" class="fz-switch-toggle" name="rss-feeds"
-												value="1" <?php echo esc_html( $disble_featured_image ); ?> />
-											<label for="rss-feeds" class="form-label"><?php echo esc_html_e( 'Do NOT add the featured image to the website\'s RSS feed.', 'feedzy-rss-feeds' ); ?></label>
-										</div>
-									</div>
-									<div class="fz-form-group">
+										<label for="feed-post-default-thumbnail" class="form-label"><?php echo esc_html_e( 'Fallback image for imported posts', 'feedzy-rss-feeds' ); ?></label>
 										<div class="help-text pb-8"><?php esc_html_e( 'Select an image to be the fallback featured image(Feed2Post).', 'feedzy-rss-feeds' ); ?></div>
 										<?php
 										$btn_label = esc_html__( 'Choose image', 'feedzy-rss-feeds' );
@@ -135,7 +126,16 @@
 											<a href="javascript:;" class="feedzy-remove-media btn btn-outline-primary <?php echo $default_thumbnail_id ? esc_attr( 'is-show' ) : ''; ?>"><?php esc_html_e( 'Remove', 'feedzy-rss-feeds' ); ?></a>
 											<input type="hidden" name="default-thumbnail-id" id="feed-post-default-thumbnail" value="<?php echo esc_attr( $default_thumbnail_id ); ?>">
 										</div>
-										<div class="help-text"><?php esc_html_e( 'This image will be used for the Feed Items that don\'t have one.', 'feedzy-rss-feeds' ); ?></div>
+										<div class="help-text">
+											<?php
+											echo wp_kses(
+												__( 'This image will be used for the <strong>imported posts</strong> if an image is not available in the source XML Feed.', 'feedzy-rss-feeds' ),
+												array(
+													'strong' => true,
+												)
+											);
+											?>
+											</div>
 									</div>
 								</div>
 								<div class="form-block">
@@ -163,7 +163,10 @@
 													<label class="form-label"><?php esc_html_e( 'First cron execution time', 'feedzy-rss-feeds' ); ?></label>
 													<input type="hidden" name="fz_execution_offset" id="fz-execution-offset" value="<?php echo ! empty( $settings['general']['fz_execution_offset'] ) ? esc_attr( $settings['general']['fz_execution_offset'] ) : ''; ?>">
 													<input type="datetime-local" id="fz-event-execution" name="fz_cron_execution" class="form-control" value="<?php echo ! empty( $settings['general']['fz_cron_execution'] ) ? esc_attr( $settings['general']['fz_cron_execution'] ) : ''; ?>">
-													<div class="help-text pt-8"><?php esc_html_e( 'When past date will be provided or left empty, event will be executed in the next queue.', 'feedzy-rss-feeds' ); ?></div>
+													<div class="help-text pt-8">
+														<?php esc_html_e( 'When past date will be provided, event will be executed in the next queue.', 'feedzy-rss-feeds' ); ?>
+														<a href="<?php echo esc_url( 'https://docs.themeisle.com/article/1820-how-to-set-scheduler-for-import-cron-jobs-in-feedzy' ); ?>" target="_blank"><?php esc_html_e( 'Learn More', 'feedzy-rss-feeds' ); ?></a>
+													</div>
 												</div>
 											</div>
 											<div class="fz-form-col-6">
@@ -197,6 +200,16 @@
 										</div>
 									</div>
 								<?php endif; ?>
+								<div class="form-block">
+									<div class="fz-form-switch pb-0">
+										<input type="checkbox" id="feedzy-telemetry" class="fz-switch-toggle" name="feedzy-telemetry"
+										value="1" <?php checked( 'yes', $telemetry_enabled ); ?> />
+										<label for="feedzy-telemetry" class="form-label"><?php esc_html_e( 'Enable Telemetry', 'feedzy-rss-feeds' ); ?></label>
+									</div>
+									<div class="fz-form-group">
+										<div class="help-text pt-8"><?php esc_html_e( 'Send data about plugin settings to measure the usage of the features. The data is private and not shared with third-party entities. Only plugin data is collected without sensitive information.' ); ?></div>
+									</div>
+								</div>
 							</div>
 							<?php
 							break;